new try

+++ hashmap.c
@@ -109,7 +109,7 @@ Pair * searchMap(HashMap * map,  char * key)
 
 Pair * firstMap(HashMap * map)
 {
-  for (long i = 0; i < map->capacity; i++)
+  for (int i = 0; i < map->capacity; i++)
     {
       if (map->buckets[i] != NULL)
       {
@@ -117,7 +117,6 @@ Pair * firstMap(HashMap * map)
         return map->buckets[i];
       }
     }
-
     return NULL;
 }
